move check_types and cast_numerics to Explorer.Shared
