+++ lib/explorer/polars_backend/series.ex
@@ -2,6 +2,7 @@ defmodule Explorer.PolarsBackend.Series do
   @moduledoc false
 
   import Kernel, except: [length: 1]
+  import Explorer.Shared, only: [check_types: 1, cast_numerics: 2]
 
   alias Explorer.DataFrame
   alias Explorer.PolarsBackend.Native
@@ -384,8 +385,8 @@ defmodule Explorer.PolarsBackend.Series do
   @impl true
   def transform(series, fun) do
     list = series |> Series.to_list() |> Enum.map(fun)
-    type = Series.check_types(list)
-    {list, type} = Series.cast_numerics(list, type)
+    type = check_types(list)
+    {list, type} = cast_numerics(list, type)
 
     from_list(list, type)
   end
